# api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
tags:
- name: Task Management
  description: APIs for managing tasks
paths:
  /api/tasks:
    post:
      tags:
      - Task Management
      summary: Create a new task
      description: Creates a new task with the provided details
      operationId: createTask
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskRequestDto"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "201":
          description: Task created successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskResponseDto"
  /api/tasks/{taskId}:
    get:
      tags:
      - Task Management
      summary: Get task by ID
      description: Retrieves a task by its ID
      operationId: getTaskById
      parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: Task retrieved successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskResponseDto"
    delete:
      tags:
      - Task Management
      summary: Delete a task
      description: Deletes a task by its ID
      operationId: deleteTask
      parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "204":
          description: Task deleted successfully
    patch:
      tags:
      - Task Management
      summary: Update an existing task
      description: Updates the details of an existing task
      operationId: updateTask
      parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdateDto"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: Task updated successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskResponseDto"
  /api/tasks/status/{status}:
    get:
      tags:
      - Task Management
      summary: Get tasks by status
      description: Retrieves tasks filtered by their status
      operationId: getTasksByStatus
      parameters:
      - name: status
        in: path
        required: true
        schema:
          type: string
          enum:
          - PENDING
          - IN_PROGRESS
          - COMPLETED
          - ON_HOLD
          - CANCELLED
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: Tasks retrieved successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskResponseDto"
  /api/tasks/priority/{priority}:
    get:
      tags:
      - Task Management
      summary: Get tasks by priority
      description: Retrieves tasks filtered by their priority
      operationId: getTasksByPriority
      parameters:
      - name: priority
        in: path
        required: true
        schema:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
          - URGENT
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: Tasks retrieved successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskResponseDto"
  /api/tasks/dueDate/{dueDate}:
    get:
      tags:
      - Task Management
      summary: Get tasks by due date
      description: Retrieves tasks filtered by their due date
      operationId: getTasksByDueDate
      parameters:
      - name: dueDate
        in: path
        required: true
        schema:
          type: string
          format: date-time
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: Tasks retrieved successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskResponseDto"
  /api/tasks/all:
    get:
      tags:
      - Task Management
      summary: Get all tasks
      description: Retrieves all tasks
      operationId: getAllTasks
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: Tasks retrieved successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskResponseDto"
components:
  schemas:
    TaskRequestDto:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        priority:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
          - URGENT
        status:
          type: string
          enum:
          - PENDING
          - IN_PROGRESS
          - COMPLETED
          - ON_HOLD
          - CANCELLED
        dueDate:
          type: string
          format: date-time
      required:
      - description
      - dueDate
      - priority
      - status
      - title
    TaskResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
          - URGENT
        status:
          type: string
          enum:
          - PENDING
          - IN_PROGRESS
          - COMPLETED
          - ON_HOLD
          - CANCELLED
        dueDate:
          type: string
          format: date-time
    TaskUpdateDto:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
          - URGENT
        status:
          type: string
          enum:
          - PENDING
          - IN_PROGRESS
          - COMPLETED
          - ON_HOLD
          - CANCELLED
        dueDate:
          type: string
          format: date-time