# src/main/java/com/hilal/TaskManagerAPI/dto/TaskRequestDto.java
package com.hilal.TaskManagerAPI.dto;

import com.hilal.TaskManagerAPI.enums.Priority;
import com.hilal.TaskManagerAPI.enums.TaskStatus;
import jakarta.persistence.*;
import jakarta.validation.constraints.*;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;
import org.springframework.format.annotation.DateTimeFormat;

import java.time.LocalDateTime;


@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class TaskRequestDto {
    @NotBlank(message = "Title cannot be blank")
    String title;

    @NotBlank(message = "Description cannot be blank")
    String description;

    @NotNull
    @Enumerated(EnumType.STRING)
    Priority priority;

    @NotNull
    @Enumerated(EnumType.STRING)
    TaskStatus status;

    @NotNull(message = "Due date cannot be null")
    @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
    LocalDateTime dueDate;
}